requirements.txt
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
python-jose==3.5.0
supabase==2.28.2
